Move ExtraConfig related func from session_util to util/configspec
