+++ pkg/vmprovider/providers/vsphere/session/session_util.go
@@ -13,32 +13,6 @@ const (
 	OvfEnvironmentTransportGuestInfo = "com.vmware.guestInfo"
 )
 
-func ExtraConfigToMap(input []vimTypes.BaseOptionValue) (output map[string]string) {
-	output = make(map[string]string)
-	for _, opt := range input {
-		if optValue := opt.GetOptionValue(); optValue != nil {
-			// Only set string type values
-			if val, ok := optValue.Value.(string); ok {
-				output[optValue.Key] = val
-			}
-		}
-	}
-	return
-}
-
-// MergeExtraConfig adds the key/value to the ExtraConfig if the key is not present, to let to the value be
-// changed by the VM. The existing usage of ExtraConfig is hard to fit in the reconciliation model.
-func MergeExtraConfig(extraConfig []vimTypes.BaseOptionValue, newMap map[string]string) []vimTypes.BaseOptionValue {
-	merged := make([]vimTypes.BaseOptionValue, 0)
-	ecMap := ExtraConfigToMap(extraConfig)
-	for k, v := range newMap {
-		if _, exists := ecMap[k]; !exists {
-			merged = append(merged, &vimTypes.OptionValue{Key: k, Value: v})
-		}
-	}
-	return merged
-}
-
 // GetMergedvAppConfigSpec prepares a vApp VmConfigSpec which will set the vmMetadata supplied key/value fields. Only
 // fields marked userConfigurable and pre-existing on the VM (ie. originated from the OVF Image)
 // will be set, and all others will be ignored.
