+++ pkg/vmprovider/providers/vsphere/session/session_vm_customization.go
@@ -192,7 +192,7 @@ func GetCloudInitGuestInfoCustSpec(
 	}
 
 	configSpec := &vimTypes.VirtualMachineConfigSpec{}
-	configSpec.ExtraConfig = MergeExtraConfig(config.ExtraConfig, extraConfig)
+	configSpec.ExtraConfig = util.MergeExtraConfig(config.ExtraConfig, extraConfig)
 
 	// Remove the VAppConfig to ensure Cloud-Init inside of the guest does not
 	// activate and prefer the OVF datasource over the VMware datasource.
@@ -217,7 +217,7 @@ func GetExtraConfigCustSpec(
 	}
 
 	configSpec := &vimTypes.VirtualMachineConfigSpec{}
-	configSpec.ExtraConfig = MergeExtraConfig(config.ExtraConfig, extraConfig)
+	configSpec.ExtraConfig = util.MergeExtraConfig(config.ExtraConfig, extraConfig)
 	return configSpec
 }
